# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Precificação de Equipamentos API",
    "description": "API para precificação de equipamentos médico-hospitalares usando busca semântica",
    "version": "2.0.0"
  },
  "paths": {
...
    "/detalhes/{grupo}": {
      "get": {
        "tags": [
          "search"
        ],
        "summary": "Detalhes Grupo",
        "description": "Retorna detalhes de um grupo de equipamentos.",
        "operationId": "detalhes_grupo_detalhes__grupo__get",
        "parameters": [
          {
            "name": "grupo",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Grupo"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/kit": {
      "get": {
        "tags": [
          "user"
        ],
        "summary": "List Kit",
        "description": "Lista os itens do kit do usuário.",
        "operationId": "list_kit_kit_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "user"
        ],
        "summary": "Add To Kit",
        "description": "Adiciona um item ao kit.",
        "operationId": "add_to_kit_kit_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/KitItem"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/kit/{kit_id}": {
      "delete": {
        "tags": [
          "user"
        ],
        "summary": "Remove From Kit",
        "description": "Remove um item do kit.",
        "operationId": "remove_from_kit_kit__kit_id__delete",
        "parameters": [
          {
            "name": "kit_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Kit Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "KitItem": {
        "properties": {
          "item_name": {
            "type": "string",
            "title": "Item Name"
          },
          "price": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Price"
          },
          "qty": {
            "type": "integer",
            "title": "Qty",
            "default": 1
          },
          "extra": {
            "anyOf": [
              {
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Extra"
          }
        },
        "type": "object",
        "required": [
          "item_name"
        ],
        "title": "KitItem",
        "description": "Modelo para adicionar item ao kit/orçamento."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}